// src/app/models/payment-profile.ts
export class PaymentProfile {
    customerProfileId: string;
    merchantCustomerId: number;
    description: string;
    email: string;
    customerType: string;
    firstName: string;
    lastName: string;
    address: string;
    city: string;
    state: string;
    zip: string;
    country: string;
    phoneNumber: string;
    cardNumber: string;
    expirationDate: string;
    defaultPaymentProfile: boolean;
    validationMode: string;
    cardCode: string;
}