// src/app/models/cardValidProfile.ts
export class CardValidProfile{
    customerProfileId: string;
    merchantCustomerId: number;
    description: string;
    email: string;
    customerType: string;
    firstName: string;
    lastName: string;
    address: string;
    city: string;
    state: string;
    zip: string;
    country: string;
    phoneNumber: string;
    cardNumber: string;
    expirationDate: string;
    defaultPaymentProfile: boolean;
    validationMode: string;
    cardCode: string;
    CardValidationAmount: number;
}